Move classes to folders
